Improve the signal processing code
When we convert from images to signals, map the pixel values to [-1,1).
Make sure we apply the background color when we draw the registration
marks. The matched filter should be 0 everywhere except where the signal
is present, which means the image should be 50% gray except where we
draw the registration mark.